fix: move biz logic to domain layer
