+++ src/webapp/components/survey/SurveyForm.tsx
@@ -21,6 +21,7 @@ import {
     PPS_PATIENT_TRACKER_TREATMENT_STAGE_ID,
 } from "../../../data/utils/surveyFormMappers";
 import { useOfflineSnackbar } from "../../hooks/useOfflineSnackbar";
+import { Questionnaire } from "../../../domain/entities/Questionnaire/Questionnaire";
 
 export interface SurveyFormProps {
     hideForm: () => void;
@@ -65,8 +66,10 @@ export const SurveyForm: React.FC<SurveyFormProps> = props => {
         props.currentSurveyId
     );
 
-    const { indicationOptions, treatmentOptions, removeLinkedStage, autoUpdateIndicationLinks } =
-        useTreatmentIndicationLink(props.formType, questionnaire);
+    const { indicationOptions, treatmentOptions, removeLinkedStage } = useTreatmentIndicationLink(
+        props.formType,
+        questionnaire
+    );
 
     useEffect(() => {
         if (saveCompleteState && saveCompleteState.status === "success") {
@@ -108,12 +111,15 @@ export const SurveyForm: React.FC<SurveyFormProps> = props => {
     const saveSurveyFormWithoutRedirect = () => {
         setLoading(true);
         if (questionnaire) {
-            const { updatedQuestionnaire, error } = autoUpdateIndicationLinks(questionnaire);
-            if (error) {
-                setLoading(false);
-                return;
-            }
-            saveSurvey(updatedQuestionnaire, true);
+            return Questionnaire.autoUpdateIndicationLinks(questionnaire).match({
+                success: updatedQuestionnaire => {
+                    saveSurvey(updatedQuestionnaire, true);
+                },
+                error: error => {
+                    snackbar.error(error.message);
+                    setLoading(false);
+                },
+            });
         }
     };
